Fixed the calendar with proper dates

+++ src/src/com/example/tson/HomeFragment.java
@@ -87,7 +87,7 @@ public class HomeFragment extends Fragment implements View.OnTouchListener
         
         int dateDifference = 0;
         try{
-	         dateDifference = bundle.getInt("dateDifference");
+	         dateDifference =(int) bundle.getLong("dateDifference");
 	         homeFragmentCalendar.add(Calendar.DAY_OF_YEAR, dateDifference);
         }catch(Exception e){Log.d("HerregudNull", "Nu blev det null!!!!");}
